feat: Add `hash_first` config option
The new default is that a file's hash appears before its name in a
directory buffer rather than after it (i.e. `hash_first = true`). This
was chosen because it makes renaming files easier (`C` clears the full
filename but keeps the hash) and prevents the hash from bouncing around
if you're adding a long filename.

If you want to keep the old behavior, then add the following to your
init.lua.

    require("dirbuf").setup {
        hash_first = false,
    }